+++ R/ggbumpribbon-package.R
@@ -1,6 +1,5 @@
 #' @keywords internal
 "_PACKAGE"
 
-#' @importFrom rlang .data
 #' @importFrom ggplot2 ggproto Stat aes %+replace%
 NULL